move callDc function to new ApiHelper.java
- refactoring in preparation of place names instead of DCID changes
